src/dpo/requirements.txt
accelerate==1.8.1
alignment_handbook==0.4.0.dev0
torch==2.5.1
#auto_gptq==0.7.1
bleach==6.2.0
boto3==1.39.12
botocore==1.39.12
bottle==0.13.4
datasets==3.6.0
deepspeed==0.15.4
gradio==5.38.1
huggingface_hub==0.33.2
hydra-core==1.3.2
jsonlines==4.0.0
multiprocess==0.70.16
numpy==2.3.1
omegaconf==2.3.0
openai==1.97.1
pandas==2.3.1
pebble==5.1.1
peft==0.16.0
Pillow==11.3.0
policy==1.0.0
PyYAML==6.0.2
PyYAML==6.0.2
qwen_vl_utils==0.0.11
regex==2024.11.6
Requests==2.32.4
rouge_score==0.1.2
safetensors==0.5.3
scipy==1.16.0
sympy==1.13.1
timeout_decorator==0.5.0
tqdm==4.67.1
transformers==4.46.3
trl==0.12.2
vllm==0.7.0
wandb==0.21.0
